requirements.txt
click==6.6
decorator==4.0.10
dj-database-url==0.4.1
Django==1.10
django-configurations==2.0
django-filter==0.13.0
django-nose==1.4.4
django-rq==0.9.1
django-versatileimagefield==1.6
djangorestframework==3.4.3
drf-nested-routers==0.11.1
ipython==5.0.0
ipython-genutils==0.1.0
Markdown==2.6.6
nose==1.3.7
pexpect==4.2.0
pickleshare==0.7.3
Pillow==3.3.0
pkg-resources==0.0.0
prompt-toolkit==1.0.5
psycopg2==2.6.1
ptyprocess==0.5.1
Pygments==2.1.3
pytz==2016.6.1
redis==2.10.5
rq==0.6.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.2
wcwidth==0.1.7
whitenoise==3.2
